% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranova-package.R
\docType{package}
\name{coranova-package}
\alias{coranova}
\alias{coranova-package}
\title{coranova: Compare Correlated Correlations}
\description{
This package can be used to formally compare two or more polygenic scores in one or more populations using correlation with outcome variable to compare the scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gunns2/coranova}
  \item Report bugs at \url{https://github.com/gunns2/coranova/issues}
}

}
\author{
\strong{Maintainer}: Sophia Gunn \email{gunns2@bu.edu} (\href{https://orcid.org/0000-0001-6451-050X}{ORCID})

}
\keyword{internal}
