% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eur}
\alias{eur}
\title{EUR data}
\format{
A dataset with 5000 rows and 3 variables:
\describe{
\item{pheno}{Simulated Outcome}
\item{pgs1}{first simulated polygenic score}
\item{pgs2}{second simulated polygenic score}
\item{pgs3}{third simulated polygenic score}
}
}
\usage{
eur
}
\description{
simulated data from an European population
}
\examples{
eur
}
\keyword{datasets}
