% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eur2000}
\alias{eur2000}
\title{EUR data 2000}
\format{
A dataset with 2000 rows and 4 variables:
\describe{
\item{pheno}{Simulated Outcome}
\item{pgs1}{first simulated polygenic score}
\item{pgs2}{second simulated polygenic score}
\item{pgs3}{third simulated polygenic score}
}
}
\usage{
eur2000
}
\description{
simulated data from an European population with 2000 individuals
}
\examples{
eur2000
}
\keyword{datasets}
