% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranova.R
\name{perform_alt_test}
\alias{perform_alt_test}
\title{Perform User-Specified Hypothesis Test}
\usage{
perform_alt_test(dat_list, outcome, measures, contrast, method, print, B)
}
\arguments{
\item{dat_list}{list of data frames, where each data frame refers to a separate population sample}

\item{outcome}{name of column containing outcome variable (must be common across dataframes in dat_list)}

\item{measures}{vector of column names of measures to be compared}

\item{contrast}{contrast matrix to generate hypothesis test}

\item{method}{to request parametric ("parametric") or bootstrap implementation ("boot") of covariance matrix V}

\item{print}{TRUE or FALSE to request print statements}

\item{B}{number of bootstraps if method == "boot" is chosen}
}
\value{
results of coranova test, when group == 1 only performs within test, when nscores = 2 provides CI
}
\description{
Perform User-Specified Hypothesis Test
}
\examples{
#TO DO
}
