% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranova.R
\name{perform_coranova}
\alias{perform_coranova}
\title{Perform Coranova}
\usage{
perform_coranova(dat_list, outcome, measures, method, B)
}
\arguments{
\item{dat_list}{list of data frames, where each data frame refers to a separate population sample}

\item{outcome}{name of outcome variable (must be common across dataframes in dat_list)}

\item{measures}{names of measures to be compared}

\item{method}{whether to use parametric or bootstrap to determine covariance
matrix of measures' correlation with outcome}

\item{B}{number of bootstrap samples}
}
\value{
results of coranova test, when group == 1 only performs within test, when nscores = 2 provides CI
}
\description{
Perform Coranova
}
\examples{
\dontrun{mat <- matrix(c(1, 0.5, 0.6, 0.5, 1, 0.2, 0.6, 0.3, 1), ncol = 3)}
\dontrun{datA <- as.data.frame(MASS::mvrnorm(n = 100,  rep(0, 3), mat))}
\dontrun{datB <- as.data.frame(MASS:mvrnorm(n = 100,  rep(0, 3), mat))}
\dontrun{a <- list(cor(datA), cor(datB))}
\dontrun{perform_coranova(a, "V1", c("V2", "V3"), "parametric")}

}
