% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranova.R
\name{perform_coranova_bootV}
\alias{perform_coranova_bootV}
\title{Perform Coranova with bootstrapped V}
\usage{
perform_coranova_bootV(dat_list, outcome, measures, B)
}
\arguments{
\item{dat_list}{list of data frames, where each data frame refers to a separate population sample}

\item{outcome}{name of outcome variable}

\item{measures}{vector of measures to be compared}

\item{B}{number of bootstrap samples to run}
}
\value{
p-value of chosen test
}
\description{
Perform Coranova with bootstrapped V
}
\examples{
 #perform_coranova_bootV(list(afr, eur), "pheno", c("pgs1", "pgs2", "pgs3"), 5, 5, "int")


}
