% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranova.R
\name{perform_coranova_parametric}
\alias{perform_coranova_parametric}
\title{Perform Coranova}
\usage{
perform_coranova_parametric(dat_list, outcome, measures)
}
\arguments{
\item{dat_list}{list of data frames, where each data frame refers to a separate population sample}

\item{outcome}{name of outcome variable (must be common across dataframes in dat_list)}

\item{measures}{names of measures to be compared}
}
\value{
results of coranova test, when group == 1 only performs within test, when nscores = 2 provides CI
}
\description{
Perform Coranova
}
\examples{
perform_coranova_parametric(list(afr, eur), "pheno", c("pgs1", "pgs2", "pgs3"))

}
