% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop2}
\alias{pop2}
\title{pop2 data}
\format{
A dataset with 20000 rows and 13 variables:
\describe{
\item{pheno}{Simulated Outcome}
\item{gwas1_methodA}{pgs developed with gwas 1 and method A}
\item{gwas2_methodA}{pgs developed with gwas 2 and method A}
\item{gwas3_methodA}{pgs developed with gwas 3 and method A}
\item{gwas4_methodA}{pgs developed with gwas 4 and method A}
\item{gwas1_methodB}{pgs developed with gwas 1 and method B}
\item{gwas2_methodB}{pgs developed with gwas 2 and method B}
\item{gwas3_methodB}{pgs developed with gwas 3 and method B}
\item{gwas4_methodB}{pgs developed with gwas 4 and method B}
\item{gwas1_methodC}{pgs developed with gwas 1 and method C}
\item{gwas2_methodC}{pgs developed with gwas 2 and method C}
\item{gwas3_methodC}{pgs developed with gwas 3 and method C}
\item{gwas4_methodC}{pgs developed with gwas 4 and method C}
}
}
\usage{
pop2
}
\description{
simulated data to use for example 2 vignette
}
\examples{
pop2
}
\keyword{datasets}
