% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{populate_mu}
\alias{populate_mu}
\title{Populate Mu}
\usage{
populate_mu(cormat_list, outcome, measures)
}
\arguments{
\item{cormat_list}{list of correlation matrices of input data frames, where each data frame refers to a separate population sample}

\item{outcome}{name of outcome variable (must be common across dataframes in dat_list)}

\item{measures}{names of measures to be compared}
}
\value{
vector of correlations according to order of measures
}
\description{
Populate Mu
}
\examples{
 #TO DO
}
