% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{populate_V}
\alias{populate_V}
\title{Populate Covariance Matrix of Correlation Values}
\usage{
populate_V(cormat_list, outcome, measures, n_list)
}
\arguments{
\item{cormat_list}{list of correlation matrices of input data frames, where each data frame refers to a separate population sample}

\item{outcome}{name of outcome variable (must be common across dataframes in dat_list)}

\item{measures}{names of measures to be compared}

\item{n_list}{list of sample sizes of input data frames}
}
\value{
covariance matrix of correlation values
}
\description{
Populate Covariance Matrix of Correlation Values
}
\examples{
#TODO
}
