% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranova-package.R
\docType{package}
\name{coranova-package}
\alias{coranova}
\alias{coranova-package}
\title{coranova: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gunns2/coranova}
  \item Report bugs at \url{https://github.com/gunns2/coranova/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
