% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{populate_R}
\alias{populate_R}
\title{Populate Sample Correlation Vector}
\usage{
populate_R(cormat_list, outcome, measures)
}
\arguments{
\item{cormat_list}{list of correlation matrices, one for each sample}

\item{outcome}{name of outcome variable}

\item{measures}{vector of names of measure variables}
}
\value{
vector of correlations with measures and outcome variable
}
\description{
Populate Sample Correlation Vector
}
\examples{
\dontrun{mat <- matrix(c(1, 0.5, 0.6, 0.5, 1, 0.2, 0.6, 0.3, 1), ncol = 3)}
\dontrun{datA <- as.data.frame(MASS::mvrnorm(n = 100,  rep(0, 3), mat))}
\dontrun{datB <- as.data.frame(MASS:mvrnorm(n = 100,  rep(0, 3), mat))}
\dontrun{a <- list(cor(datA), cor(datB))}
\dontrun{populate_R(a, "V1", c("V2", "V3"))}
}
